+++ CHANGELOG.md
@@ -4,6 +4,33 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.7-preview.5] - 2019-12-10
+
+### Changed
+* Updated `com.unity.platforms` to `0.1.7-preview.3`.
+
+## [0.1.7-preview.4] - 2019-11-21
+
+### Changed
+* Updated `Run` to open output folder, for now
+
+## [0.1.7-preview.3] - 2019-11-12
+
+### Changed
+* Made fields public on `DotsWebTarget`
+* Added Compatibility with new dots runtime build customization framework
+* Removed old bad coupling with dots repo
+
+## [0.1.7-preview.2] - 2019-11-04
+
+### Changed
+* Added build settings component for emscripten settings. 
+
+## [0.1.7-preview] - 2019-10-25
+
+### Changed
+* Updated `com.unity.platforms` to `0.1.7-preview`.
+
 ## [0.1.6-preview] - 2019-10-23
 
 ### Added